Calculate daily infection values

This function calculates the daily infection values for Cercospora beticola on sugar beet. Functions were adapted from Wolf and Verreet (2005) and Wolf et al (2001)

Usage

calc_DIV(date_time, Tm, RH, rain, dat)

Arguments

date_time

POSIX_ct, date time the weather recording was taken

Tm

numeric, temperature, in Celsius' at time increment in date_time

RH

numeric, relative humidity (%) at time increment in date_time

rain

numeric, volume of rain in millimetres recorded between time recordings

dat

data.frame, containing column names "times","temp","rh","rain" with each of the respective arguments for input. provided as a convenience

Value

data.table, with probability of infection for each day, between 0 and 1 Undertaken with two methods by Wolf and Verreet (2005)

References

Wolf, P. F. J., and J. A. Verreet. “Factors Affecting the Onset of Cercospora Leaf Spot Epidemics in Sugar Beet and Establishment of Disease-Monitoring Thresholds.” Phytopathology® 95, no. 3 (March 2005): 269–74. https://doi.org/10.1094/PHYTO-95-0269.

Wolf, P. F. J., M. Heindl, and J. A. Verreet. “Influence of Sugar Beet Leaf Mass Development on Predisposition of the Crop to Cercospora Beticola (Sacc.).” Journal of Plant Diseases and Protection 108, no. 6 (2001): 578–92.

Examples

date_t <- Sys.time() + seq(0, 179 * 60 * 10, (60 * 10))
Tm <- rnorm(180,20,10)
RH <- runif(180,min = 40,90)
rain <- rbinom(180,1,0.1) * runif(180,0.1,20)

DIV1 <- calc_DIV(
date_time = date_t,
Tm = Tm,
RH = RH,
rain = rain
)