Read LAI images at several time points

Read LAI images at several time points

Usage

read_sb_growth_parameter(im_folder, target_res)

Arguments

im_folder

folder containing two georeferenced image of the location of interest collected at two time points. The two images should be named after the date of collection (eg. 2022-05-01.tif)

target_res

desired spatial resolution. target_res should be equal to or larger than the actual resolution of the images expressed in meters.

Value

param_list: list containing parameters that are necessary to calculate r, the growth rate.

Examples

epidemic_onset_param <-
  read_sb_growth_parameter(im_folder = system.file("extdata", "uav_img",
                                                   package = "cercospoRa"),
                           target_res = 10)
epidemic_onset_param$tm
#> [1] "2022-06-14 UTC" "2022-06-28 UTC"
epidemic_onset_param$imgs
#> class       : SpatRaster 
#> dimensions  : 17, 29, 2  (nrow, ncol, nlyr)
#> resolution  : 0.0001103147, 6.885404e-05  (x, y)
#> extent      : 9.915826, 9.919025, 51.49783, 51.499  (xmin, xmax, ymin, ymax)
#> coord. ref. : lon/lat WGS 84 (EPSG:4326) 
#> source(s)   : memory
#> names       : 2022_06_14, 2022_06_28 
#> min values  :  0.5216454,  0.8607956 
#> max values  :  2.9568495,  3.3260000