Calculate daily infection values

This function calculates the daily infection values for Cercospora berticola on sugar beet. Functions were adapted from (Wolf and Verreet 2005) and (Wolf et al. 2001)

Usage

calc_DIV(date_time, Tm, RH, rain, dat)

Arguments

date_time

POSIX_ct, date time the weather recording was taken

Tm

numeric, temperature, in celcius' at time increment in date_time

RH

numeric, relative humidity (%) at time increment in date_time

rain

numeric, volume of rain in millimeters recorded between time recordings

dat

data.frame, containing column names "times","temp","rh","rain" with each of the respective arguments for input. provided as a convenience

Value

data.table, with probability of infection for each day, between 0 and 1 Undertaken with two methods by Wolf (Wolf and Verreet 2005)

under the DIV column and method by Racca et. al (Racca and Jörg 2007)

described in the DIV_racca column.

References

Racca P, Jörg E (2007). “CERCBET 3 – a forecaster for epidemic development of Cercospora beticola*.” EPPO Bulletin, 37(2), 344--349. ISSN 1365-2338, doi:10.1111/j.1365-2338.2007.01136.x , _ eprint: https://onlinelibrary.wiley.com/doi/pdf/10.1111/j.1365-2338.2007.01136.x, https://onlinelibrary.wiley.com/doi/abs/10.1111/j.1365-2338.2007.01136.x.

Wolf PFJ, Heindl M, Verreet JA (2001). “Zum Einfluss des Bestandesklimas auf die Prädisposition der Zuckerrübe gegenüber Infektionen von Cercospora beticola (Sacc.) / Influence of sugar beet leaf mass development on predisposition of the crop to Cercospora beticola (Sacc.).” Zeitschrift für Pflanzenkrankheiten und Pflanzenschutz / Journal of Plant Diseases and Protection, 108(6), 578--592. ISSN 0340-8159, Publisher: Verlag Eugen Ulmer KG, https://www.jstor.org/stable/43215404.

Wolf PFJ, Verreet JA (2005). “Factors Affecting the Onset of Cercospora Leaf Spot Epidemics in Sugar Beet and Establishment of Disease-Monitoring Thresholds.” Phytopathology®, 95(3), 269--274. ISSN 0031-949X, doi:10.1094/PHYTO-95-0269 , Publisher: Scientific Societies, https://apsjournals.apsnet.org/doi/abs/10.1094/PHYTO-95-0269.

Examples

date_t <- Sys.time() + seq(0, 179 * 60 * 10, (60 * 10))
Tm <- rnorm(180,20,10)
RH <- runif(180,min = 40,90)
rain <- rbinom(180,1,0.1) * runif(180,0.1,20)

DIV1 <- calc_DIV(
date_time = date_t,
Tm = Tm,
RH = RH,
rain = rain
)