Calculate Cercospora latent period

Calculate Cercospora latent period

Usage

latent_period(Tm, ref = "wolf")

Arguments

Tm

numeric, Average temperature in degrees Celsius for a hour of the day

ref

character, method for calculating latent period. Default is "wolf" also available "jarroudi". See references for where formulas were used

Value

numeric, Latent period for a given hour. ref = "jarroudi" returns latent period in days.

Details

Calculates the latent period for Cercospora beticola infections on sugar beet

References

Wolf PFJ, Verreet JA (2005). “Factors Affecting the Onset of Cercospora Leaf Spot Epidemics in Sugar Beet and Establishment of Disease-Monitoring Thresholds.” Phytopathology®, 95(3), 269--274. ISSN 0031-949X, doi:10.1094/PHYTO-95-0269 , Publisher: Scientific Societies, https://apsjournals.apsnet.org/doi/abs/10.1094/PHYTO-95-0269. ; equation 4 El Jarroudi M, Chairi F, Kouadio L, Antoons K, Sallah AM, Fettweis X (2021). “Weather-Based Predictive Modeling of Cercospora beticola Infection Events in Sugar Beet in Belgium.” Journal of Fungi, 7(9), 777. ISSN 2309-608X, doi:10.3390/jof7090777 , https://www.ncbi.nlm.nih.gov/pmc/articles/PMC8470031/. ; equation 1