Get temperature index

Temperature index is a proportional representation of the latent period. Temperatures at or above the optimum temperature for the disease cycle will yield 1 and the lower the temperature the infinitely longer the latent period.

Usage

temperature_index(Tm, opt_Tm = 21)

Arguments

Tm

numeric, temperature for any given hour

opt_Tm

numeric, the lowest temperature optimum at which all temperatures above will have the same (fastest) latent period.

Value

numeric, proportion representing the speed of the latent period in relation to the temperature optimum

Details

For equations and original documentation, see (Wolf and Verreet 2005) and (Wolf et al. 2001)

References

Wolf PFJ, Heindl M, Verreet JA (2001). “Zum Einfluss des Bestandesklimas auf die Prädisposition der Zuckerrübe gegenüber Infektionen von Cercospora beticola (Sacc.) / Influence of sugar beet leaf mass development on predisposition of the crop to Cercospora beticola (Sacc.).” Zeitschrift für Pflanzenkrankheiten und Pflanzenschutz / Journal of Plant Diseases and Protection, 108(6), 578--592. ISSN 0340-8159, Publisher: Verlag Eugen Ulmer KG, https://www.jstor.org/stable/43215404.

Wolf PFJ, Verreet JA (2005). “Factors Affecting the Onset of Cercospora Leaf Spot Epidemics in Sugar Beet and Establishment of Disease-Monitoring Thresholds.” Phytopathology®, 95(3), 269--274. ISSN 0031-949X, doi:10.1094/PHYTO-95-0269 , Publisher: Scientific Societies, https://apsjournals.apsnet.org/doi/abs/10.1094/PHYTO-95-0269.